filegramplugin/requirements.txt
watchdog>=4.0.0
pyyaml>=6.0
